% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_text}
\alias{clean_text}
\title{Function to remove some html tags function from the supreme court data}
\usage{
clean_text(court_data, col_of_interest)
}
\arguments{
\item{court_data}{A data frame containing details about supreme court cases.}

\item{col_of_interest}{A column of the `court_data` data frame that you wish to clean.}
}
\value{
A data frame
}
\description{
Function to remove some html tags function from the supreme court data
}
\details{
Some column cells have some unnecessary characters in the beginning or end of
}
\examples{
clean_text(court_data,
           href)
}
